refactor: cafeteria 타입 최적화
