+++ src/interfaces/Cafeteria.ts
@@ -1,8 +1,30 @@
+export type MealType = 'BREAKFAST' | 'LUNCH' | 'DINNER';
+
+export type MealName = '아침' | '점심' | '저녁';
+
+export type MealTypes = Array<'BREAKFAST' | 'LUNCH' | 'DINNER'>;
+
+export type PlaceType = 'A코너' | 'B코너' | 'C코너' | '능수관' | '2캠퍼스';
+
+export const placeOrder = ['A코너', 'B코너', 'C코너', '능수관', '2캠퍼스'];
+
+export const MEAL_TYPE_MAP = {
+  BREAKFAST: '아침',
+  LUNCH: '점심',
+  DINNER: '저녁',
+} as const;
+
+export type CafeteriaTime = {
+  id: number;
+  type: MealType;
+  name: MealName;
+}[];
+
 export type CafeteriaMenu = {
   id: number;
   date: string;
-  type: 'BREAKFAST' | 'LUNCH' | 'DINNER';
-  place: 'A코너' | 'B코너' | 'C코너' | '능수관' | '2캠퍼스';
+  type: MealType;
+  place: PlaceType;
   price_card: number | null;
   price_cash: number | null;
   kcal: number | null;
@@ -13,14 +35,3 @@ export type CafeteriaMenu = {
   soldout_at: string | null;
   changed_at: string | null;
 };
-
-export type MealType = 'BREAKFAST' | 'LUNCH' | 'DINNER';
-export type MealName = '아침' | '점심' | '저녁';
-
-export type MealTypes = Array<'BREAKFAST' | 'LUNCH' | 'DINNER'>;
-
-export type CafeteriaTime = {
-  id: number;
-  type: MealType;
-  name: MealName;
-}[];
