+++ src/pages/IndexPage/components/IndexCafeteria/index.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable jsx-a11y/click-events-have-key-events */
 import { useNavigate } from 'react-router-dom';
-import { CAFETERIA_CATEGORY } from 'static/cafeteria';
+import { CafeteriaType, CAFETERIA_CATEGORY } from 'static/cafeteria';
 import { useState } from 'react';
 import { ReactComponent as RightArrow } from 'assets/svg/right-arrow.svg';
 import { cn } from '@bcsdlab/utils';
